requirements.txt
discord.py==1.6.0
Pillow==7.2.0
requests==2.25.1
urllib3==1.26.3
